fix np.float

+++ evaluator/ava_evaluation/metrics.py
@@ -94,7 +94,7 @@ def compute_average_precision(precision, recall):
 
     if not isinstance(precision, np.ndarray) or not isinstance(recall, np.ndarray):
         raise ValueError("precision and recall must be numpy array")
-    if precision.dtype != np.float or recall.dtype != np.float:
+    if precision.dtype != np.float32 or recall.dtype != np.float32:
         raise ValueError("input must be float numpy array.")
     if len(precision) != len(recall):
         raise ValueError("precision and recall must be of the same size.")
